Finalize sidebar key typing
